move addUTMToLinks() to helpers file
